src/mumble/CoreAudio: Reformatted files.
